+++ Source/ACE/Command/Handlers/DebugCommands.cs
@@ -621,5 +621,68 @@ namespace ACE.Command.Handlers
             var positionMessage = new GameMessageSystemChat(message, ChatMessageType.Broadcast);
             session.Network.EnqueueSend(positionMessage);
         }
+
+        /// <summary>
+        /// Debug command to learn a spell.
+        /// </summary>
+        /// <param name="parameters">A single uint spell id within between 1 and 6340. (Not all spell ids are valid.)</param>
+        [CommandHandler("learnspell", AccessLevel.Developer, CommandHandlerFlag.RequiresWorld, 0,
+            "[(uint)spellid] - Adds the specificed spell to your spellbook (non-persistant).",
+            "@learnspell")]
+        public static void HandleLearnSpell(Session session, params string[] parameters)
+        {
+            if (parameters?.Length > 0)
+            {
+                uint spellId = (uint)int.Parse(parameters[0]);
+
+                SpellTable spells = SpellTable.ReadFromDat();
+                if (!spells.Spells.ContainsKey(spellId))
+                {
+                    var errorMessage = new GameMessageSystemChat("SpellID not found in Spell Table", ChatMessageType.Broadcast);
+                    session.Network.EnqueueSend(errorMessage);
+                }
+                else
+                {
+                    var updateSpellEvent = new GameEventMagicUpdateSpell(session, spellId);
+                    session.Network.EnqueueSend(updateSpellEvent);
+
+                    // Always seems to be this SkillUpPurple effect
+                    session.Player.ActionApplyVisualEffect(PlayScript.SkillUpPurple, session.Player.Guid);
+
+                    string spellName = spells.Spells[spellId].Name;
+                    // TODO Lookup the spell in the spell table.
+                    string message = "You learn the " + spellName + " spell.\n";
+                    var learnMessage = new GameMessageSystemChat(message, ChatMessageType.Broadcast);
+                    session.Network.EnqueueSend(learnMessage);
+                }
+            }
+            else
+            {
+                string message = "Invalid Syntax\n";
+                var errorMessage = new GameMessageSystemChat(message, ChatMessageType.Broadcast);
+                session.Network.EnqueueSend(errorMessage);
+            }
+        }
+
+        /// <summary>
+        /// Debug command to print out all of the active players connected too the server.
+        /// </summary>
+        [CommandHandler("listplayers", AccessLevel.Developer, CommandHandlerFlag.RequiresWorld, 0,
+            "Displays all of the active players connected too the serve.",
+            "@players")]
+        public static void HandleListPlayers(Session session, params string[] parameters)
+        {
+            uint playerCounter = 0;
+            // Build a string message containing all available characters and send as a System Chat message
+            string message = "";
+            foreach (Session playerSession in WorldManager.GetAll(false))
+            {
+                message += $"{playerSession.Player.Name} : {(uint)playerSession.Id}\n";
+                playerCounter++;
+            }
+            message += $"Total connected Players: {playerCounter}\n";
+            var listPlayersMessage = new GameMessageSystemChat(message, ChatMessageType.Broadcast);
+            session.Network.EnqueueSend(listPlayersMessage);
+        }
     }
 }